Don't make the arraysize() macro cast to int.
Reviewed-on: https://code-review.googlesource.com/c/re2/+/43030
(cherry picked from commit 7e9d7987aabfd14198f1380de8e0057d8992ac05)

+++ re2/testing/compile_test.cc
@@ -113,7 +113,7 @@ static Test tests[] = {
 
 TEST(TestRegexpCompileToProg, Simple) {
   int failed = 0;
-  for (int i = 0; i < arraysize(tests); i++) {
+  for (size_t i = 0; i < arraysize(tests); i++) {
     const re2::Test& t = tests[i];
     Regexp* re = Regexp::Parse(t.regexp, Regexp::PerlX|Regexp::Latin1, NULL);
     if (re == NULL) {
